+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.1.0 29.01.2026
 ## Added
 
-- Added Function "Create-Project"
+- Added Function "New-Project"
 - Added Function "Set-ProjectRoot"
 
 ## Changed